Fix annoying SteamVR pop-up after making a change
A SteamVR window always pops up on an unmodified copy of the SteamVR
plugin. Since it is impossible to disable to Unity splash screen on the
free version, this window will pop up every time a script is edited or
Play mode is started. The change listed in this commit needs to be made
every time the SteamVR plugin is updated.