Add preamble UI and ability to pause experiment
A UI will appear before the participant/experimenter is able to start the experiment. This will allow them to perform any necessary
calibrations as well as the ability to pause the experiment at the start
of each trial during the fixation checking stage to perform any
necessary re-calibrations or re-evalutations. The functionality of the
menu items will be added one-by-one in future commits.